Add mock BFD9010 scanner service and Docker support
- Implement mock BFD9010 service with FHIR Device endpoints in `mock_bfd9010.py`
- Create Dockerfile for building the mock service
- Update Makefile to include build target for mock BFD9010 image
- Modify docker-compose.yml to configure mock service and expose necessary ports
- Add sample images for testing in `sample_images.bz2`
