+++ Makefile
@@ -1,8 +1,9 @@
 .DEFAULT_GOAL := help
 
 COMPOSE_FILE ?= bfd9000_web/docker-compose.yml
+MOCK_BFD9010_IMAGE ?= mock-bfd9010:local
 
-.PHONY: help build-test-docker run-test-docker mock-bfd9010
+.PHONY: help build-test-docker run-test-docker build-mock-bfd9010 mock-bfd9010
 
 help: ## Show available targets
 	@awk 'BEGIN {FS = ":.*##"} /^[a-zA-Z0-9_-]+:.*##/ {printf "%-20s %s\n", $$1, $$2}' $(MAKEFILE_LIST)
@@ -13,5 +14,8 @@ build-test-docker: ## Build test Docker image with compose
 run-test-docker: ## Run test Docker stack with compose
 	docker compose -f $(COMPOSE_FILE) up
 
+build-mock-bfd9010: ## Build mock BFD9010 Docker image
+	docker build -t $(MOCK_BFD9010_IMAGE) dev_tools/mock_bfd9010
+
 mock-bfd9010: ## Start the mock BFD9010 scanner service on port 5000
-	python dev_tools/mock_bfd9010.py --host localhost
+	python dev_tools/mock_bfd9010/mock_bfd9010.py --host localhost
